Generalize OpenMM converters
