+++ .github/workflows/ci.yaml
@@ -28,7 +28,7 @@ jobs:
         make docs-build
 
     - name: CodeCov
-      uses: codecov/codecov-action@v3.1.1
+      uses: codecov/codecov-action@v4.1.1
       with:
         file: ./coverage.xml
         flags: unittests
