+++ smee/__init__.py
@@ -5,6 +5,7 @@ Differentiably evaluate energies of molecules using SMIRNOFF force fields
 """
 
 from . import _version
+from ._constants import EnergyFn, PotentialType
 from ._models import (
     NonbondedParameterMap,
     ParameterMap,
@@ -23,6 +24,8 @@ from .potentials import compute_energy, compute_energy_potential
 __version__ = _version.get_versions()["version"]
 
 __all__ = [
+    "EnergyFn",
+    "PotentialType",
     "ValenceParameterMap",
     "NonbondedParameterMap",
     "ParameterMap",
