+++ smee/converters/openmm/__init__.py
@@ -1,0 +1,17 @@
+"""Convert tensor representations into OpenMM systems."""
+
+from smee.converters.openmm._openmm import (
+    convert_to_openmm_force,
+    convert_to_openmm_system,
+    convert_to_openmm_topology,
+    create_openmm_system,
+    potential_converter,
+)
+
+__all__ = [
+    "convert_to_openmm_force",
+    "convert_to_openmm_system",
+    "convert_to_openmm_topology",
+    "create_openmm_system",
+    "potential_converter",
+]
